For tutor dashboard in mobile view, sidebar menu is not there for users to navigate.
make changes to make sure we have sidebar navigation in mobile view also, dont make any other changes like import of icons and all because it it creating issues which you are not able to solve, stick to the scope of this requirement. this is only for tutor dashboard having sidebar navigation in mobile view also. dont make changes any where else if needed to create a new layout specifically for tutor dashboard we can do it.